refactor: separate resolution kinds from emit input module
