+++ crates/syntax/src/program/mod.rs
@@ -2,10 +2,10 @@ mod definition;
 mod emit_input;
 mod file;
 mod module;
+mod resolution;
 
 pub use definition::{Definition, Interface, MethodSignatures, Visibility};
-pub use emit_input::{
-    CallKind, DotAccessKind, EmitInput, MutationInfo, NativeTypeKind, ReceiverCoercion, UnusedInfo,
-};
+pub use emit_input::{EmitInput, MutationInfo, UnusedInfo};
 pub use file::{File, FileImport};
 pub use module::{Module, ModuleId, ModuleInfo};
+pub use resolution::{CallKind, DotAccessKind, NativeTypeKind, ReceiverCoercion};
